fix: precision loss in RangeVal.plus and NumVal.compareTo
RangeVal.plus: replace toInt() with Double comparison preserving
original NumVal instances. NumVal.compareTo: replace core.toInt()
with core.toDouble() to avoid Long/Float truncation. Unsure.new:
replace else branch with explicit is NullVal for sealed exhaustiveness.

+++ commons-value/src/main/kotlin/edu/jhu/cobra/commons/value/primitive/NumVal.kt
@@ -241,5 +241,5 @@ data class NumVal(override val core: Number) : IPrimitiveVal {
      * @param other The integer to compare with.
      * @return A comparison result.
      */
-    operator fun compareTo(other: Int): Int = core.toInt().compareTo(other)
+    operator fun compareTo(other: Int): Int = core.toDouble().compareTo(other.toDouble())
 }
